feat(geth): use measurement mode branch
Since we want to only make one type of measurements at the time and we
want to have fair and unaffected measurements of the execution,
we decided to prepare separate branches for each individual measurement
mode.
In this scenario, each branch differs from the `master` as little
as possible and consists only of changes that are required for the specific
measurement type.